+++ src/app.ts
@@ -1,394 +1,6 @@
-// Drag & Drop Interfaces
-interface Draggable {
-  dragStartHandler(e: DragEvent): void;
-  dragEndHandler(e: DragEvent): void;
-}
+import { ProjectInput } from './components/project-input.js';
+import { ProjectList } from './components/project-list.js';
 
-interface DragTarget {
-  dragOverHandler(e: DragEvent): void;
-  dropHandler(e: DragEvent): void;
-  dragLeaveHandler(e: DragEvent): void;
-}
-
-enum ProjectStatus {
-  Active,
-  Finished,
-}
-
-// Project Type
-class Project {
-  constructor(
-    public id: string,
-    public title: string,
-    public description: string,
-    public people: number,
-    public status: ProjectStatus
-  ) {}
-}
-
-// Project State Management class
-type Listener<T> = (items: T[]) => void;
-
-class State<T> {
-  protected listeners: Listener<T>[] = [];
-
-  addListener(listenerFunction: Listener<T>) {
-    this.listeners.push(listenerFunction);
-  }
-}
-
-class ProjectState extends State<Project> {
-  private projects: Project[] = [];
-  private static instance: ProjectState;
-
-  private constructor() {
-    super();
-  }
-
-  static getInstance() {
-    if (this.instance) {
-      return this.instance;
-    }
-    this.instance = new ProjectState();
-    return this.instance;
-  }
-
-  addProject(title: string, description: string, numOfPeople: number) {
-    // Math.random for testing only
-    const id = Math.random().toString();
-    const newProject = new Project(
-      id,
-      title,
-      description,
-      numOfPeople,
-      ProjectStatus.Active
-    );
-
-    this.projects.push(newProject);
-    for (const listenerFunction of this.listeners) {
-      listenerFunction(this.projects.slice());
-    }
-  }
-
-  moveProject(projectId: string, newStatus: ProjectStatus) {
-    const project = this.projects.find(({ id }) => id === projectId);
-    if (project && project.status !== newStatus) {
-      project.status = newStatus;
-      this.updateListeners();
-    }
-  }
-
-  private updateListeners() {
-    for (const listenerFunction of this.listeners) {
-      listenerFunction(this.projects.slice());
-    }
-  }
-}
-
-const projectState = ProjectState.getInstance();
-
-// Validation
-interface Validatable {
-  value: string | number;
-  required?: boolean;
-  minLength?: number;
-  maxLength?: number;
-  min?: number;
-  max?: number;
-}
-
-function validate(validatableInput: Validatable) {
-  let isValid = true;
-
-  if (validatableInput.required) {
-    isValid = isValid && validatableInput.value.toString().trim().length !== 0;
-  }
-  if (
-    validatableInput.minLength != null &&
-    typeof validatableInput.value === 'string'
-  ) {
-    isValid =
-      isValid &&
-      validatableInput.value.trim().length >= validatableInput.minLength;
-  }
-  if (
-    validatableInput.maxLength != null &&
-    typeof validatableInput.value === 'string'
-  ) {
-    isValid =
-      isValid &&
-      validatableInput.value.trim().length <= validatableInput.maxLength;
-  }
-  if (
-    validatableInput.min != null &&
-    typeof validatableInput.value === 'number'
-  ) {
-    isValid = isValid && validatableInput.value >= validatableInput.min;
-  }
-  if (
-    validatableInput.max != null &&
-    typeof validatableInput.value === 'number'
-  ) {
-    isValid = isValid && validatableInput.value <= validatableInput.max;
-  }
-
-  return isValid;
-}
-
-// autobind decorator
-const autoBind = (
-  _target: any,
-  _method: string,
-  descriptor: PropertyDescriptor
-) => {
-  const originalMethod = descriptor.value;
-  const adjDescriptor: PropertyDescriptor = {
-    configurable: true,
-    get() {
-      const boundFunction = originalMethod.bind(this);
-      return boundFunction;
-    },
-  };
-  return adjDescriptor;
-};
-
-// Component Base Class
-abstract class Component<T extends HTMLElement, U extends HTMLElement> {
-  templateElement: HTMLTemplateElement;
-  hostElement: T;
-  element: U;
-
-  constructor(
-    templateId: string,
-    hostElementId: string,
-    insertAtStart: boolean,
-    newElementId?: string
-  ) {
-    this.templateElement = document.getElementById(
-      templateId
-    )! as HTMLTemplateElement;
-    this.hostElement = document.getElementById(hostElementId)! as T;
-
-    const importedNode = document.importNode(
-      this.templateElement.content,
-      true
-    );
-    this.element = importedNode.firstElementChild as U;
-    if (newElementId) {
-      this.element.id = newElementId;
-    }
-
-    this.attach(insertAtStart);
-  }
-
-  private attach(insertAtBeginning: boolean) {
-    this.hostElement.insertAdjacentElement(
-      insertAtBeginning ? 'afterbegin' : 'beforeend',
-      this.element
-    );
-  }
-
-  abstract configure(): void;
-  abstract renderContent(): void;
-}
-
-// ProjectInput Class
-class ProjectInput extends Component<HTMLDivElement, HTMLFormElement> {
-  titleInputElement: HTMLInputElement;
-  descriptionInputElement: HTMLInputElement;
-  peopleInputElement: HTMLInputElement;
-
-  constructor() {
-    super('project-input', 'app', true, 'user-input');
-
-    this.titleInputElement = this.element.querySelector(
-      '#title'
-    ) as HTMLInputElement;
-    this.descriptionInputElement = this.element.querySelector(
-      '#description'
-    ) as HTMLInputElement;
-    this.peopleInputElement = this.element.querySelector(
-      '#people'
-    ) as HTMLInputElement;
-
-    this.configure();
-  }
-
-  configure() {
-    this.element.addEventListener('submit', this.submitHandler);
-  }
-
-  renderContent() {}
-
-  private gatherUserInput(): [string, string, number] | void {
-    const enteredTitle = this.titleInputElement.value;
-    const enteredDescription = this.descriptionInputElement.value;
-    const enteredPeople = this.peopleInputElement.value;
-
-    const titleValidatable = {
-      value: enteredTitle,
-      required: true,
-    };
-    const descriptionValidatable = {
-      value: enteredDescription,
-      required: true,
-      minLength: 5,
-    };
-    const peopleValidatable = {
-      value: +enteredPeople,
-      required: true,
-      min: 1,
-      max: 10,
-    };
-
-    const isValid =
-      validate(titleValidatable) &&
-      validate(descriptionValidatable) &&
-      validate(peopleValidatable);
-
-    if (!isValid) {
-      alert('Invalid input');
-      return;
-    } else {
-      return [enteredTitle, enteredDescription, +enteredPeople];
-    }
-  }
-
-  private clearInputs() {
-    this.titleInputElement.value = '';
-    this.descriptionInputElement.value = '';
-    this.peopleInputElement.value = '';
-  }
-
-  @autoBind
-  private submitHandler(e: Event) {
-    e.preventDefault();
-    const userInput = this.gatherUserInput();
-    if (Array.isArray(userInput)) {
-      const [title, description, people] = userInput;
-      projectState.addProject(title, description, people);
-      this.clearInputs();
-    }
-  }
-}
-
-// Project Item Class
-class ProjectItem
-  extends Component<HTMLUListElement, HTMLLIElement>
-  implements Draggable {
-  private project: Project;
-
-  get persons() {
-    return this.project.people === 1
-      ? '1 person'
-      : `${this.project.people} persons`;
-  }
-
-  constructor(hostId: string, project: Project) {
-    super('single-project', hostId, true, project.id);
-    this.project = project;
-
-    this.configure();
-    this.renderContent();
-  }
-
-  @autoBind
-  dragStartHandler(e: DragEvent) {
-    e.dataTransfer!.setData('text/plain', this.project.id);
-    e.dataTransfer!.effectAllowed = 'move';
-  }
-
-  @autoBind
-  dragEndHandler(_e: DragEvent) {
-    console.log('Drag End');
-  }
-
-  configure() {
-    this.element.addEventListener('dragstart', this.dragStartHandler);
-    this.element.addEventListener('dragend', this.dragEndHandler);
-  }
-
-  renderContent() {
-    this.element.querySelector('h2')!.textContent = this.project.title;
-    this.element.querySelector('h3')!.textContent = this.persons + ' assigned';
-    this.element.querySelector('p')!.textContent = this.project.description;
-  }
-}
-
-// ProjectList Class
-class ProjectList
-  extends Component<HTMLDivElement, HTMLElement>
-  implements DragTarget {
-  assignedProjects: Project[];
-
-  constructor(private type: 'active' | 'finished') {
-    super('project-list', 'app', false, `${type}-projects`);
-    this.assignedProjects = [];
-
-    this.configure();
-    this.renderContent();
-  }
-
-  @autoBind
-  dragOverHandler(e: DragEvent) {
-    if (e.dataTransfer && e.dataTransfer.types[0] === 'text/plain') {
-      e.preventDefault();
-      const listElement = this.element.querySelector('ul')!;
-      listElement.classList.add('droppable');
-    }
-  }
-
-  @autoBind
-  dropHandler(e: DragEvent) {
-    const projectId = e.dataTransfer!.getData('text/plain');
-    projectState.moveProject(
-      projectId,
-      this.type === 'active' ? ProjectStatus.Active : ProjectStatus.Finished
-    );
-  }
-
-  @autoBind
-  dragLeaveHandler(_e: DragEvent) {
-    const listElement = this.element.querySelector('ul')!;
-    listElement.classList.remove('droppable');
-  }
-
-  configure() {
-    this.element.addEventListener('dragover', this.dragOverHandler);
-    this.element.addEventListener('dragleave', this.dragLeaveHandler);
-    this.element.addEventListener('drop', this.dropHandler);
-
-    projectState.addListener((projects: Project[]) => {
-      const relativeProjects = projects.filter(({ status }) => {
-        if (this.type === 'active') {
-          return status === ProjectStatus.Active;
-        }
-        return status === ProjectStatus.Finished;
-      });
-      this.assignedProjects = relativeProjects;
-      this.renderProjects();
-    });
-  }
-
-  renderContent() {
-    const listId = `${this.type}-projects-list`;
-    this.element.querySelector('ul')!.id = listId;
-    this.element.querySelector('h2')!.textContent =
-      this.type.toUpperCase() + ' PROJECTS';
-  }
-
-  private renderProjects() {
-    const listElement = document.getElementById(
-      `${this.type}-projects-list`
-    )! as HTMLUListElement;
-    listElement.innerHTML = '';
-
-    for (const project of this.assignedProjects) {
-      new ProjectItem(this.element.querySelector('ul')!.id, project);
-    }
-  }
-}
-
-const projectInput = new ProjectInput();
-const activeList = new ProjectList('active');
-const finishedList = new ProjectList('finished');
+new ProjectInput();
+new ProjectList('active');
+new ProjectList('finished');
