+++ src/components/project-list.ts
@@ -1,0 +1,80 @@
+import { Component } from '../components/base-component.js';
+import { DragTarget } from '../models/drag-drop.js';
+import { autoBind } from '../decorators/autobind.js';
+import { Project, ProjectStatus } from '../models/project.js';
+import { projectState } from '../state/project-state.js';
+import { ProjectItem } from '../components/project-item.js';
+
+// ProjectList Class
+export class ProjectList
+  extends Component<HTMLDivElement, HTMLElement>
+  implements DragTarget {
+  assignedProjects: Project[];
+
+  constructor(private type: 'active' | 'finished') {
+    super('project-list', 'app', false, `${type}-projects`);
+    this.assignedProjects = [];
+
+    this.configure();
+    this.renderContent();
+  }
+
+  @autoBind
+  dragOverHandler(e: DragEvent) {
+    if (e.dataTransfer && e.dataTransfer.types[0] === 'text/plain') {
+      e.preventDefault();
+      const listElement = this.element.querySelector('ul')!;
+      listElement.classList.add('droppable');
+    }
+  }
+
+  @autoBind
+  dropHandler(e: DragEvent) {
+    const projectId = e.dataTransfer!.getData('text/plain');
+    projectState.moveProject(
+      projectId,
+      this.type === 'active' ? ProjectStatus.Active : ProjectStatus.Finished
+    );
+  }
+
+  @autoBind
+  dragLeaveHandler(_e: DragEvent) {
+    const listElement = this.element.querySelector('ul')!;
+    listElement.classList.remove('droppable');
+  }
+
+  configure() {
+    this.element.addEventListener('dragover', this.dragOverHandler);
+    this.element.addEventListener('dragleave', this.dragLeaveHandler);
+    this.element.addEventListener('drop', this.dropHandler);
+
+    projectState.addListener((projects: Project[]) => {
+      const relativeProjects = projects.filter(({ status }) => {
+        if (this.type === 'active') {
+          return status === ProjectStatus.Active;
+        }
+        return status === ProjectStatus.Finished;
+      });
+      this.assignedProjects = relativeProjects;
+      this.renderProjects();
+    });
+  }
+
+  renderContent() {
+    const listId = `${this.type}-projects-list`;
+    this.element.querySelector('ul')!.id = listId;
+    this.element.querySelector('h2')!.textContent =
+      this.type.toUpperCase() + ' PROJECTS';
+  }
+
+  private renderProjects() {
+    const listElement = document.getElementById(
+      `${this.type}-projects-list`
+    )! as HTMLUListElement;
+    listElement.innerHTML = '';
+
+    for (const project of this.assignedProjects) {
+      new ProjectItem(this.element.querySelector('ul')!.id, project);
+    }
+  }
+}
