+++ src/decorators/autobind.ts
@@ -1,0 +1,15 @@
+export const autoBind = (
+  _target: any,
+  _method: string,
+  descriptor: PropertyDescriptor
+) => {
+  const originalMethod = descriptor.value;
+  const adjDescriptor: PropertyDescriptor = {
+    configurable: true,
+    get() {
+      const boundFunction = originalMethod.bind(this);
+      return boundFunction;
+    },
+  };
+  return adjDescriptor;
+};
